chore: update proxies/models/resources to align with data-service v0.3.2 (#290)
Fixing the enum issue.  The problem here is there seems to be an implict use of `enum#valueOf` which cannot be overwritten.  Unless we change the case of our enums to lowercase, to make this work we need to have all capitals.  The use of `String#toUpperCase` here gives us a bit of leniency on the API with capital letters, however, it does feel a little cheesy as we'll have to remember to use toUpperCase on any new enum usage.