fix: bad document id given at load of application
Adding safety code here to handle `undefined` being passed in as a String value. There are a few related issues here like "why can this asset not be found" and "why have we hardcoded to look at the second document in the list and not the first" but those can be addressed later.