removed navigation panel from Dataset, Model and Document pages
I removed the table of contents with links to scroll to different sections of the page. It occupied too much space.

fyi @shawn I discussed your idea of showing/hiding that feature depending on window width with @pascaleproulx and we decided to remove it (for now) because we felt the content switcher should be on the right side of the title. We'll revisit the scroll-to-section feature down the road.