feat: new "Related publications" component
Added a component near the tops of the Dataset and Model pages for users to select which resources they want to use to enrich a model or dataset.

I blocked in the UI for this. I need someone to populate the dialog with a list of documents and datasets in the project, and then actually send the selected items to the enrichment service.