feat(hmi-server): add endpoint to see if name already exists
Adding a new endpoint to the server to check if an asset already exists with that name. This can optionally be done within the scope of a specific project or across all entries of that asset type.  In addition, blowing away old unused `toString()` functionality inside of AssetType as it was a nightmare to work with and unused.