fix(hmi): 500 error on new models
When a new model is created, there is a `watch` call to `/ascet-to-latex` which results in a 500 error because the model is empty.  Arguably this call shouldn't be happening at all in the empty state, but, we should at least handle the empty case better. Thats whats happening here - we short circuit the request and return a 204.