fix(hmi): 405 error when posting provenance
Fixing a 405 error when posting to connected nodes. There were a few problems here. The first being that the URL was incorrect. But then the second issue was another front-end type that was out of sync with the back end, so I've blown that away and made the back end type a tsmodel to use.