fix[hmi-server]: duplicate assets can be added to a project
Our project asset controller now throws a 409 error code when a duplicate asset tries to get added to a project.

In addition, a few other fixes:
* Cleaning up unused methods in `ProjectAssetService`.
* Found a previously reported bug in our front end for adding an xdd document to a project. This was a 2-for-1 bug where the code didn't actually work as explained, and there was some typing issues that made my skin crawl :)

Also to note, the current implementation of `useProjects` and `ProjectService` hide away any helpful error messages from the back end. We may want to revisit this design in the near future. I've updated a TODO doc with this idea?