fix[hmi-server]: incorrect parsing of some pdf urls
We don't actually need to do a full encode here as most of the URLS given are coming from unpaywall and are good to go. What we mostly need to do is just replace whitespace with `%20`. This feels a little hacky to do but thats cowboy coding, baby!
This works for random spot tests on PDFs as well as the previously problematic article with a DOI of 10.1101/2020.07.30.20164921