feats[hmi-client]: The Documents / Models / Dataset switcher on the Explorer page ought to show number of items in each
This is a bit of a 'first pass' at this task because of the horribly disjointed state the explorer is in. Some of the searches (models and xdd documents) happen all at once so we can populate them.  Others (datasets) happen on demand. We also have two types of documents which makes this muddy (and in code there's only one type and then seperate checks??). As a result you won't get 100% accurate results in some of the cases until you click on the tabs which isn't great but its the best that can be done without doing the "big task" of addressing this data explorer...