Fixing wrong model configuration selection
After creating a new model configuration, we weren't properly selecting it from the the list of model configs. As a result the output and the download would be the last selected/created config even though it *looked* like it shouldn't be.
In addition I've found more casting exceptions because of the presence of `Object` in the back end. This should be `Number` but that could break our python services down stream, and requires a migration and I just want to get it in.
Lastly, we should return the new model we've created ;)