[FEAT]: Export Model Config and Model at once, zipped #4316
When a user downloads a model configuration now they will receive a zipped archive that includes the original model and the configuration.  This will preserve all the metadata and information that was included in the configuration and not present it as a 'default configuration'.  I've opted to not use the extension `.zip` because it would be slightly harder to verify whats coming in to us.