[BUG]: Better extraction failure notifications
There are two fixes here. Firstly, for #5154 we now send up an error message for the front end with a comma seperated list of what extraction services failed. Secondly for #5123 we forgot to send a *Final* message in the failure case so the extraction just sat forever at its last reported %

Resolves #5123 and #5154