Extract geolocation information when enriching model AMR
This is a first pass at getting geoname information. Right now we need to fall back to doing the DKG search on MIRA but in the future we could probably also host this information locally?
I envision the search for geonames potentially happening seperately from regular DKG searches, with the locations being merged into the standard DKG search if needed.