[FEAT]: Copy model interventions and configs with model #5739
Manually checking for a model to be copied, then check for intervention policies or configurations associated with that model to clone. The manual check has to be done because currently we walk the json of the `model` in a one-direction way, but models don't contain references to their configs/interventions... thats the other way around. This was a quicker fix for a deadline. Better fix would be to have models include references outward and then this problem goes away.